requirements.txt
alembic==1.13.1
barnum==0.5.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==0.18.3
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
passlib==1.7.4
python-dotenv==1.0.1
setuptools==69.0.3
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
